vehicle fetching on app-root level, reload on oracle switch
