+++ web/src/elements/app-root.ts
@@ -1,13 +1,26 @@
 import { LitElement, html } from 'lit';
-import { customElement } from 'lit/decorators.js';
+import {customElement, state} from 'lit/decorators.js';
 import { provide } from '@lit/context';
 import { apiServiceContext } from '../context';
 import {ApiService} from "@services/api-service.ts";
+import {Vehicle} from "@datatypes/vehicle.ts";
+
+interface VehiclesResponse {
+    vehicles: Vehicle[];
+}
 
 @customElement('app-root')
 export class AppRoot extends LitElement {
     @provide({ context: apiServiceContext })
     apiService = ApiService.getInstance(); // app-level singleton
+
+    @state()
+    private vehicles: Vehicle[]
+
+    constructor() {
+        super();
+        this.vehicles = []
+    }
 
     // enable inherit css
     createRenderRoot() {
@@ -24,17 +37,28 @@ export class AppRoot extends LitElement {
             <oracle-selector @option-changed=${this.handleOracleChange}></oracle-selector>
             <add-vin-element></add-vin-element>
             
-            <vehicle-list-element></vehicle-list-element>
+            <vehicle-list-element .items=${this.vehicles}></vehicle-list-element>
     `;
     }
 
-    private handleOracleChange(e: CustomEvent) {
+    private async handleOracleChange(e: CustomEvent) {
         // Access the selected value from the event detail
         const selectedValue = e.detail.value;
         console.log('Oracle changed to:', selectedValue);
 
         this.apiService.setOracle(selectedValue);
-        // todo: reload vehicle-list
+        await this.getUserVehicles()
     }
 
+    async connectedCallback() {
+        super.connectedCallback();
+        await this.getUserVehicles()
+    }
+
+    async getUserVehicles() {
+        const url = "/vehicles";
+        const userVehiclesResponse = await this.apiService.callApi<VehiclesResponse>('GET', url, null, true);
+        console.debug('user vehicles', userVehiclesResponse)
+        this.vehicles = userVehiclesResponse.data?.vehicles || [];
+    }
 }
