+++ web/src/elements/vehicle-list-element.ts
@@ -1,13 +1,7 @@
 import {html, LitElement} from 'lit'
 import {repeat} from 'lit/directives/repeat.js';
-import {ApiService} from "@services/api-service.ts";
-import {customElement} from "lit/decorators.js";
+import {customElement, property} from "lit/decorators.js";
 import {Vehicle} from "@datatypes//vehicle.ts";
-
-
-interface VehiclesResponse {
-    vehicles: Vehicle[];
-}
 
 @customElement('vehicle-list-element')
 export class VehicleListElement extends LitElement {
@@ -15,14 +9,17 @@ export class VehicleListElement extends LitElement {
         items: {type: Array},
         alertText: {type: String },
     }
+
+    @property({attribute: true})
     private items: Vehicle[];
+
     private alertText: string;
-    private api: ApiService;
+
 
     constructor() {
         super();
         this.items = [];
-        this.api = ApiService.getInstance();
+
         this.alertText = "";
     }
 
@@ -30,17 +27,6 @@ export class VehicleListElement extends LitElement {
     createRenderRoot() {
         // there is another function to do this.
         return this;
-    }
-
-    async connectedCallback() {
-        super.connectedCallback();
-        // call func to load items
-        const userVehiclesResponse = await this.getUserVehicles();
-        if(!userVehiclesResponse.success) {
-            this.alertText = "failed to get vehicles: " + userVehiclesResponse.error;
-        }
-        console.debug('user vehicles', userVehiclesResponse)
-        this.items = userVehiclesResponse.data?.vehicles || [];
     }
 
     render() {
@@ -61,23 +47,5 @@ export class VehicleListElement extends LitElement {
           <vehicle-list-item-element .item="${item}">`)}
             </table>
         `
-    }
-
-    async getUserVehicles() {
-        const url = "/vehicles";
-        return await this.api.callApi<VehiclesResponse>('GET', url, null, true);
-    }
-
-    async disconnectVehicle(vin: string) {
-        const url = "/vehicle/disconnect";
-        const body = {vin};
-
-        return await this.api.callApi<any>('POST', url, body, true);
-    }
-
-    async deleteVehicle(vin: string) {
-        const url = "/vehicle/delete";
-        const body = {vin};
-        return await this.api.callApi<any>('POST', url, body, true);
     }
 }
